move more methods from `DiscreteGeometry` to `Mesh`
